+++ plugins/packaging-tasks/src/functionalTest/groovy/com/apgsga/packaging/rpm/tasks/test/CopyResourcesToBuildDirActionTests.groovy
@@ -1,7 +1,7 @@
 package com.apgsga.packaging.rpm.tasks.test
 
 import com.apgsga.gradle.test.utils.AbstractSpecification
-import com.apgsga.packaging.plugins.ApgPackaging
+import com.apgsga.packaging.plugins.ApgRpmPackagePlugin
 
 class CopyResourcesToBuildDirActionTests extends AbstractSpecification {
 
@@ -9,7 +9,7 @@ class CopyResourcesToBuildDirActionTests extends AbstractSpecification {
         given:
         buildFile << """
             plugins {
-                id '${ApgPackaging.PLUGIN_ID}' 
+                id '${ApgRpmPackagePlugin.PLUGIN_ID}' 
             }
         """
