--- plugins/packaging-tasks/src/main/resources/META-INF/gradle-plugins/com.apgsga.packaging.properties
@@ -1,1 +1,0 @@
-implementation-class=com.apgsga.packaging.plugins.ApgPackaging
