+++ plugins/packaging-tasks/src/main/resources/META-INF/gradle-plugins/com.apgsga.service.package.properties
@@ -1,0 +1,1 @@
+implementation-class=com.apgsga.packaging.plugins.ApgServicePackagePlugin
