+++ plugins/version-manager/src/main/kotlin/com/apgsga/maven/dm/plugin/VersionResolutionPlugin.kt
@@ -2,7 +2,7 @@ package com.apgsga.maven.dm.plugin
 
 import com.apgsga.gradle.repo.plugin.ApgCommonRepoPlugin
 import com.apgsga.maven.dm.ext.VersionResolutionExtension
-import com.apgsga.packaging.plugins.ApgPackaging
+import com.apgsga.packaging.plugins.ApgCommonPackagePlugin
 import com.apgsga.revision.manager.domain.RevisionManagerBuilder
 import org.gradle.api.Action
 import org.gradle.api.Plugin
@@ -15,7 +15,7 @@ open class VersionResolutionPlugin : Plugin<Project> {
     override fun apply(project: Project) {
         project.plugins.apply(ApgCommonRepoPlugin::class.java)
         project.plugins.apply(MavenPublishPlugin::class.java)
-        project.plugins.apply(ApgPackaging::class.java)
+        project.plugins.apply(ApgCommonPackagePlugin::class.java)
         val revisionManagerBuilder = RevisionManagerBuilder.create()
         val extension =  project.extensions.create("apgVersionResolver", VersionResolutionExtension::class.java, project, revisionManagerBuilder)
         applyRecommendations(project, extension)
